feat: moves llmo controller to its own folder along with its mappings file
